refactor: move migrations to different file
